Versi 1.61

+++ Clinic_dev/Clinic/Payment/BillList.cs
@@ -1186,7 +1186,7 @@ namespace Clinic
             }
 
 
-            if (Convert.ToInt16(txtLimitStok.Text) <= 0)
+            if (Convert.ToInt32(txtLimitStok.Text) <= 0)
             {
                 limit = "5";
                 txtLimitStok.Text = "5";
@@ -1448,19 +1448,19 @@ namespace Clinic
 
                 if (stok != "")
                 {
-                    if (Convert.ToInt16(stok) == 0)
+                    if (Convert.ToInt32(stok) == 0)
                     {
                         e.Appearance.BackColor = Color.Crimson;
                         e.Appearance.ForeColor = Color.White;
                         e.Appearance.FontStyleDelta = FontStyle.Bold;
                     }
-                    else if (Convert.ToInt16(stok) <= 20)
+                    else if (Convert.ToInt32(stok) <= 20)
                     {
                         e.Appearance.BackColor = Color.FromArgb(150, Color.OrangeRed);
                         e.Appearance.ForeColor = Color.White;
                         e.Appearance.FontStyleDelta = FontStyle.Bold;
                     }
-                    else if (Convert.ToInt16(stok) > 20)
+                    else if (Convert.ToInt32(stok) > 20)
                     {
                         e.Appearance.BackColor = Color.FromArgb(150, Color.Green);
                         e.Appearance.ForeColor = Color.White;
@@ -1575,13 +1575,13 @@ namespace Clinic
 
                 if (stok != "")
                 {
-                    if (Convert.ToInt16(stok) <= 0)
+                    if (Convert.ToInt32(stok) <= 0)
                     {
                         e.Appearance.BackColor = Color.Crimson;
                         e.Appearance.ForeColor = Color.White;
                         e.Appearance.FontStyleDelta = FontStyle.Bold;
                     }
-                    else if (Convert.ToInt16(stok) <= 20)
+                    else if (Convert.ToInt32(stok) <= 20)
                     {
                         e.Appearance.BackColor = Color.FromArgb(150, Color.OrangeRed);
                         e.Appearance.ForeColor = Color.White;
@@ -2268,7 +2268,7 @@ namespace Clinic
                     tmp_amt_med = "0";
                 }
 
-                if (comboBox1.Text == "Rawat Inap" && (Convert.ToInt16(tmp_diag) == 0))
+                if (comboBox1.Text == "Rawat Inap" && (Convert.ToInt32(tmp_diag) == 0))
                 {
                     MessageBox.Show("Silahkan Isi Diagnosa Akhir");
                     return;
